Split out TranslationMemoryEntry & TranslationMemoryEntryQuerySet
